requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-writable-nested==0.7.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
